feat: Add Header component for navigation and branding
feat: Implement HomePageLanding component with mentor and session features

feat: Create ScrollToTop component for smooth scrolling on route changes

feat: Develop UserProfile page to display mentor details, services, and reviews
